pyproject.toml
[project]
name = "django-sample-todos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "markdown>=3.7",
]
